fix: Correct rms_norm_eps type hint from int to float in MistralConfig
The rms_norm_eps parameter was incorrectly typed as int | None but
defaults to 1e-6 which is a float. This parameter is passed to
MistralRMSNorm which expects eps: float. This is a type hint bug
that affects 59+ model configurations in the codebase.

+++ src/transformers/models/mistral/configuration_mistral.py
@@ -69,7 +69,7 @@ class MistralConfig(PreTrainedConfig):
         hidden_act: str | None = "silu",
         max_position_embeddings: int | None = 4096 * 32,
         initializer_range: float | None = 0.02,
-        rms_norm_eps: int | None = 1e-6,
+        rms_norm_eps: float | None = 1e-6,
         use_cache: bool | None = True,
         pad_token_id: int | None = None,
         bos_token_id: int | None = 1,
